Refactor for mongodb
- Create mongodb model
- Create mongodb repositories
- Update Product, Review, Comment controllers
- Update tests
