+++ src/main/java/com/udacity/course3/reviews/controller/ReviewsController.java
@@ -1,14 +1,12 @@
 package com.udacity.course3.reviews.controller;
 
-import com.udacity.course3.reviews.model.Product;
-import com.udacity.course3.reviews.model.Review;
-import com.udacity.course3.reviews.repository.ProductRepo;
-import com.udacity.course3.reviews.repository.ReviewRepo;
-import org.apache.coyote.Response;
+import com.udacity.course3.reviews.model.mongodb.Product;
+import com.udacity.course3.reviews.model.mongodb.Review;
+import com.udacity.course3.reviews.repository.mongodb.ProductRepo;
+import com.udacity.course3.reviews.repository.mongodb.ReviewRepo;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.client.HttpServerErrorException;
 
 import java.util.List;
 import java.util.Optional;
@@ -59,6 +57,6 @@ public class ReviewsController {
     @RequestMapping(value = "/reviews/products/{productId}", method = RequestMethod.GET)
     public ResponseEntity<List<?>> listReviewsForProduct(@PathVariable("productId") Integer productId) {
         return new ResponseEntity<>(
-                reviewRepo.findByProductId(productId), HttpStatus.OK);
+                productRepo.findReviewById(productId), HttpStatus.OK);
     }
 }
