--- src/main/java/com/udacity/course3/reviews/repository/CommentRepo.java
@@ -1,11 +1,0 @@
-package com.udacity.course3.reviews.repository;
-
-import com.udacity.course3.reviews.model.Comment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface CommentRepo extends JpaRepository<Comment,Integer> {
-
-    Optional<Comment> findByIdAndComment(Integer id, String comment);
-}
