--- src/main/java/com/udacity/course3/reviews/repository/ProductRepo.java
@@ -1,11 +1,0 @@
-package com.udacity.course3.reviews.repository;
-
-import com.udacity.course3.reviews.model.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface ProductRepo extends JpaRepository<Product,Integer> {
-
-    Optional<Product> findByIdAndName(Integer id, String name);
-}
