+++ src/main/resources/application.properties
@@ -1,9 +1,15 @@
 # FIXME: Follow instructions to setup MySQL and add the following properties
-spring.datasource.url=jdbc:mysql://localhost:3306/udacity?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
-spring.datasource.username=udacity
-spring.datasource.password=udacity123
-spring.flyway.baselineOnMigrate=true
-spring.flyway.baseline-on-migrate=true
+#spring.datasource.url=jdbc:mysql://localhost:3306/udacity?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
+#spring.datasource.username=udacity
+#spring.datasource.password=udacity123
+
+spring.datasource.driver-class-name=org.h2.Driver
+spring.datasource.url=jdbc:h2:mem:localhost;DB_CLOSE_ON_EXIT=FALSE
+datasource.username=admin
+datasource.password=
+spring.data.mongodb.uri=mongodb://192.168.99.100:32768/reviews-app
+#spring.flyway.baselineOnMigrate=true
+#spring.flyway.baseline-on-migrate=true
 
 spring.jpa.show-sql=true
