+++ src/test/java/com/udacity/course3/reviews/ReviewsApplicationTests.java
@@ -7,7 +7,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
-@DataJpaTest
+//@DataJpaTest
 public class ReviewsApplicationTests {
 
 	@Test
